pyproject.toml
[tool.poetry]
name = "active-person"
version = "0.1.0"
description = "It's my pet-project. You can see more about it in miro url, which I added in Readme"
authors = ["Vitaly <lildid32@mail.ru>"]
readme = "README.md"

packages = [{ include = "server" }]

package-mode = true

[tool.poetry.scripts]
start = 'server.main:start'

[tool.pytest.ini_options]
asyncio_mode = "auto"

pythonpath = [
    "."
]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["all"], version = "^0.115.0"}
asyncpg = "^0.29.0"
sqlalchemy = "^2.0.35"
pydantic = "^2.9.2"
black = "^24.8.0"
pydantic-extra-types = "^2.9.0"
pycryptodomex = "^3.20.0"
crypto = "^1.4.1"
phonenumbers = "^8.13.46"
alembic = "^1.13.3"
python-dotenv = "^1.0.1"
cryptography = "^43.0.1"
python-jose = "^3.3.0"
redis = "^5.1.0"
celery = {extras = ["redis"], version = "^5.4.0"}
aredis = "^1.1.8"
dictutils = "^0.1.7"
passlib = "^1.7.4"
fastapi-users = {extras = ["sqlalchemy"], version = "^13.0.0"}
pytest = "^8.3.3"
pytest-asyncio = {extras = ["all"], version = "^0.24.0"}
anyio = "^4.6.2.post1"
pytest-tornasync = "^0.6.0.post2"
pytest-trio = "^0.8.0"
pytest-twisted = "^1.14.3"
twisted = "^24.7.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
